requirements.txt
matplotlib==3.3.2
transformers==3.4.0
numpy==1.19.2
pandas==1.1.2
scipy==1.5.2
six==1.15.0
onnxruntime==1.4.0
torch==1.6.0
absl_py==0.10.0
datasets==1.5.0
seaborn==0.11.0
nltk==3.5
tqdm==4.49.0
absl-py==0.10.0
rouge_score==0.0.4
